% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{calculate_sdf}
\alias{calculate_sdf}
\title{Function to calculate an SDF from a set of points to a polygon}
\usage{
calculate_sdf(points_sf, range_sf, geo_dist = TRUE)
}
\arguments{
\item{geo_dist}{}
}
\value{

}
\description{
Function to calculate an SDF from a set of points to a polygon
}
