% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{collect_sdf_samples}
\alias{collect_sdf_samples}
\title{Function to generate SDF samples for a range polygon within a background polygon.}
\usage{
collect_sdf_samples(
  range_polygons,
  bg_polygons,
  env = NULL,
  n_pts = 1e+05,
  close_scale = 0.025,
  very_close_scale = 0.0025,
  geo_dist = FALSE,
  centrer = NULL,
  scaler = NULL
)
}
\arguments{
\item{n_sdf_samples}{}
}
\value{

}
\description{
Function to generate SDF samples for a range polygon within a background polygon.
}
