% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{deepSDF_model}
\alias{deepSDF_model}
\title{Function to generate deepSDF Keras model}
\usage{
deepSDF_model(input_len, net_breadth = 256L, dropout_rate = 0.5)
}
\arguments{
\item{dropout_rate}{}
}
\value{

}
\description{
Function to generate deepSDF Keras model
}
