% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_to_iso}
\alias{df_to_iso}
\title{Function to convert a data frame with x, y, z variables into an isoline or isoband object}
\usage{
df_to_iso(
  df,
  levs,
  levs_high = NULL,
  x = "X",
  y = "Y",
  z = "sdf",
  buffer = NULL
)
}
\arguments{
\item{z}{}
}
\value{

}
\description{
Function to convert a data frame with x, y, z variables into an isoline or isoband object
}
