% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{make_cross_validations}
\alias{make_cross_validations}
\title{Function to generate validation data}
\usage{
make_cross_validations(
  sdf_samples,
  validation_type,
  validation_folds,
  validation_split,
  use_coords
)
}
\arguments{
\item{validation_split}{}
}
\value{

}
\description{
Function to generate validation data
}
