% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{make_one_cross_validation}
\alias{make_one_cross_validation}
\title{Make one cross validation from a set of intervals}
\usage{
make_one_cross_validation(sdf_samples, starts, ends, use_coords)
}
\arguments{
\item{use_coords}{}
}
\value{

}
\description{
Make one cross validation from a set of intervals
}
