% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{make_prediction_grid}
\alias{make_prediction_grid}
\title{Function to generate a grid across a background polygon on which to make model predictions}
\usage{
make_prediction_grid(
  range_sf,
  bg,
  env = NULL,
  n_cells = 10000,
  return_type = c("sf", "tibble")
)
}
\arguments{
\item{return_type}{}
}
\value{

}
\description{
Function to generate a grid across a background polygon on which to make model predictions
}
