% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{make_prediction_sf}
\alias{make_prediction_sf}
\title{Function to generate a grid across a background polygon on which to make model predictions as an sf object}
\usage{
make_prediction_sf(range_sf, bg, n_cells = 10000)
}
\arguments{
\item{geo_dist}{}
}
\value{

}
\description{
Function to generate a grid across a background polygon on which to make model predictions as an sf object
}
