% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{make_SDF_grid}
\alias{make_SDF_grid}
\title{Function to generate an SDF on a background grid for a polygon}
\usage{
make_SDF_grid(range_sf, bg, geo_dist = TRUE)
}
\arguments{
\item{geo_dist}{}
}
\value{

}
\description{
Function to generate an SDF on a background grid for a polygon
}
