% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_training_animation}
\alias{make_training_animation}
\title{Title}
\usage{
make_training_animation(
  model_fit,
  use_future = FALSE,
  freeze_frames = 0.05,
  duration = 10
)
}
\arguments{
\item{freeze_frames}{}
}
\value{

}
\description{
Title
}
