% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{plot_preds}
\alias{plot_preds}
\title{Prediction plotter}
\usage{
plot_preds(
  predictions,
  prediction_df,
  bg,
  centrer = NULL,
  scaler = NULL,
  file_names
)
}
\arguments{
\item{prediction_df}{}
}
\value{

}
\description{
Prediction plotter
}
