% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_SDF}
\alias{plot_SDF}
\title{Function to plot SDF of a polygon}
\usage{
plot_SDF(
  sdf,
  range_sf = NULL,
  bg = NULL,
  type = c("grid", "isolines"),
  inner_lines = 5
)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Function to plot SDF of a polygon
}
