% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_to_sf}
\alias{predict_to_sf}
\title{Function to make predictions from the model and return an sf object}
\usage{
predict_to_sf(model, test_dat, prediction_df)
}
\arguments{
\item{prediction_df}{}
}
\value{

}
\description{
Function to make predictions from the model and return an sf object
}
