% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SRM.R
\name{run_SRM}
\alias{run_SRM}
\title{Title}
\usage{
run_SRM(
  range_polygons,
  env_raster,
  bg_polygons = NULL,
  n_sdf_samples = 1e+05,
  batch_size = floor(n_sdf_samples/20),
  net_breadth = 256L,
  dropout_rate = 0.5,
  use_coords = FALSE,
  standardise_env = TRUE,
  vis_training_progress = TRUE,
  validation_type = c("folds", "random", "none"),
  validation_folds = 5L,
  validation_split = 4L,
  validation_prop = 0.2,
  epochs = "auto",
  keep_sdf_samples = TRUE,
  max_epochs = 500L,
  geo_dist = FALSE
)
}
\arguments{
\item{range_polygons}{A polygon or set of polygons as an sf object. The specifies the range boundaries that
are being modelled (the response).}

\item{env_raster}{A \code{RasterLayer} or \code{RasterStack} object with environmental covariates (the predictors).
The projection should match the projection of \code{range_polygons}}

\item{bg_polygons}{The projection should match the projection of \code{range_polygons}}

\item{max_epochs}{}
}
\value{

}
\description{
Title
}
