% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sf_points_to_tibble}
\alias{sf_points_to_tibble}
\title{Function to convert a points based sf to a tibble}
\usage{
sf_points_to_tibble(points_sf)
}
\arguments{
\item{points_sf}{}
}
\value{

}
\description{
Function to convert a points based sf to a tibble
}
