% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{run_model}
\alias{run_model}
\title{Function to run the model on a set of data}
\usage{
run_model(
  train_dat,
  validation_dat = NULL,
  test_dat = NULL,
  epochs,
  batch_size
)
}
\arguments{
\item{batch_size}{}
}
\value{

}
\description{
Function to run the model on a set of data
}
